Split save/autosave to NewPage
